+++ front/app/src/Router/Hooks/useSavePrevPage.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable max-statements */
-import { useAppDispatch, useAppSelector } from "../../Redux/hooks/redux-hooks";
+import { useAppDispatch } from "../../Redux/hooks/redux-hooks";
 import { setPreviousPage } from "../../Redux/store/controllerAction";
 
 export const	useSavePrevPage = (pageToSave: string) =>
