+++ front/app/src/custom.d.ts
@@ -3,3 +3,9 @@ declare module "*.svg"
 	const content: string;
 	export default content;
 }
+
+declare module "*.jpg"
+{
+	const content: string;
+	export default content;
+}
